feat(lib): VPrimitive:  conditional headless element vi asTemplate prop
Conditonal logic was being repeated on every element that could render a default element of be renderless via "as" prop. This will dry the views using a single unified api for all.

Note that this does not allow polymorphism, as in provided elements via slot scope will have to manually bind things, theres not prop/attrs/listerners merging. This is intended decision but i might give it a second thought.