fix/feat!: retrieving dome element ref universally
in a renderless this.$el no longer points to the first sibling as vue2. With the support of fragments it points to a placeholder element (comment), used to keep track of position. using `nextElementSibling` keeps the behaviour aligned with v2, but it's a bit fragile.

For now i'm relying on querySelector if the consumer has binded `attrs` provided via slot to a specific element in the case of multiple elements in the slot scope.

BREAKING CHANGE
- namespaced data attributes to avoid conflicts with other packages